eliminated separate test class for max tests
