+++ src/test/java/org/cicirello/permutations/distance/CyclicEdgeDistanceTests.java
@@ -31,6 +31,16 @@ import org.cicirello.permutations.*;
 public class CyclicEdgeDistanceTests extends SharedTestForPermutationDistance {
 	
 	@Test
+	public void testMax() {
+		CyclicEdgeDistance d = new CyclicEdgeDistance();
+		for (int n = 0; n <= 7; n++) {
+			int expected = bruteForceComputeMax(d,n);
+			assertEquals(expected, d.max(n), "Failed on length: " + n);
+			assertEquals(1.0*expected, d.maxf(n), "Failed on length: " + n);
+		}
+	}
+	
+	@Test
 	public void testIdenticalPermutations() {
 		CyclicEdgeDistance d = new CyclicEdgeDistance();
 		identicalPermutations(d);
