+++ src/test/java/org/cicirello/permutations/distance/SharedTestForPermutationDistanceDouble.java
@@ -38,4 +38,13 @@ public class SharedTestForPermutationDistanceDouble {
 			assertEquals(0.0, d.distancef(p, p), "distance of a permutation to itself should be 0; length was " + n);
 		}
 	}
+	
+	final double bruteForceComputeMaxD(PermutationDistanceMeasurerDouble d, int n) {
+		double max = 0;
+		Permutation p1 = new Permutation(n, 0);
+		for (Permutation p2 : p1) {
+			max = Math.max(max, d.distancef(p1,p2));
+		}
+		return max;
+	}
 }
